box: Add configurable random seed
This axes the last global of dash and adds it to the box class.
While at it consolidate on drand48() useage.
Use int64_t instead of long it to keep cpplint google style happy. (;

The random seed can be both set on cli and params file.
Now that the physics is right it makes sense to get
statics of different runs.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>