box: optimise collision list
We currently only need a list of integer for the collision table.
No point in copying again there the particledata as the full particle
list is accessed anyway for the collision partner.
Nevertheless was a good test / usecase for the ParticleData class.

While at it try to make collide_particles() more readable.

On a quick test this gave 10-20% performance gain.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>