box: add timespec field for startup time + reorder output
We want to know how long we used to compute,
so output it all along and at the end of the simulation.
Assign the variable directly in the constructor,
as this kind of a special case (;.

Make the mash simulation output aligned more to the left now.
At later times we might want to axe some of the outputs.
While at it bump version.

It needs special link argument for older glibc.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>