include: [FourVector.h] add relational operators
The class should follow the design principle of the least surprise
and hence directly allow such comparisons. Use the minor operator
as base for the others comparison operator. This is possible if
both left and right side are FourVector class.

This is useful for example to know if a particle is inside the box.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>