cmake: s/box/mash/ and build it in current builddir
Change date versioning to US format without leading zeros.
The year version is more important as the month or day.

Use consistently the name of the box, makes much for fun to
run it. A box could be anything this is mash!!!

"A mass of mixed ingredients reduced to a soft pulpy state
by beating or pressure"
http://en.wiktionary.org/wiki/mash

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>