box, ic, particles, outputroutines: Add particle_type map
This allows to get rid of hardcoded use of a specific particle type.
It maps the id of the particle to an existing particle type.

On quick test didn't see much performance change.

The c++ map seems both fancied by Hannah and Jochen. (:
Jochen demonstrated a specific implementation. Here by using
just a pair of int's we don't need to sort, but in fact we
don't care and could use unordered_map, but this would already
entail a c++ 2011 dependency, which is not widely supported on
available clusters yet.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>