box, IC, particles, outputroutines: Use vector particles container
This is another merge from Jussi, as this is more c++ style conformant.
Needs to have (*object)[i] instead of the more readble object[i] of a
c style array.

On the bonus side one can axe the number paramenter to many functions
as the vector container knows it's size.

While at it changed some output files to const reference.
(This should have been a seperate change but discovered it on the go).

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>