collisions: Keep a counter for each scattering that happened
This allows to know which particle just interacted,
without interferring with future calculations.

It is also deemed necessary in future for decay or particle
creation, to know which process created the particle.
This counter will count over many different processes.

As a sideeffect it excludes for good the interaction of
particles that just collided as this value will not be
overwritten while calculating the upcoming collision.

Suggested-by: Kai Gallmeister <Gallmeister@th.physik.uni-frankfurt.de>
Suggested-by: Hannah Petersen <petersen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>