IC, particles: generalize/fix momentum sampling
As long as we use the back-to-back initialization, we can sample the
cos(theta) just between 0 and 1. However, I find this a bit dangerous,
as if we happen to change the initialization in the future, we might be
sampling just half of the momentum space if we still use interval 0 to 1
instead of -1 to 1. So I would change this part of the code to give
cos(theta) a random value between -1 and 1.

phi: sampled between 0 and 2*Pi <- ok!
theta: sin(theta) sampled between 0 and 1 <- puzzled! Should be cos(theta)?

The reason I'd use cos(theta) here is because (apologies if I'm stating
obvious stuff here, just want to be as clear about my reasoning as
possible) the differential "volume" element is

d^3 p = dp_x * dp_y * dp_z = p^2 dp sin(theta) d(theta) d(phi).

So, for sampling the "radial momentum" p, we sample the value between 0
and infinity with a weight function p^2 (+ Maxwell-Boltzmann
distribution exp(-E/T) in the box).

For phi, we sample from 0 to 2*Pi with weight function 1.

For theta, we sample, in principle, from 0 to Pi with weight function
sin(theta). But as we don't really want to do the sampling from a weight
function if we can avoid it, we use a trick:

Integral of sin(theta) d(theta) from 0 to Pi
= - Integral of d[cos(theta)] from 1 to -1
= Integral of d[cos(theta)] from -1 to 1

So sampling theta between 0 and Pi with weight sin(theta) = sampling
cos(theta) between -1 and 1 with weight 1.

Fix it too on the random angle particle exchange and bump version.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>