mash: Allow configurable particle types
On startup read a particles.txt, which defines the particle types
data fields.

This needs particle types to be a std::vector too as
particle data. This is a useful first step for adding other
particles. Also this is kind of a merge from Jussi as
he had configurable particle list, but allows also to adjust
the properties of the particles currently only name, restmass,
pdgnumber.

An example of such a particles.txt would be:
pi-     0.13957         -211
pi+     0.13957         211
pi0     0.134977        111
rho     0.775           113

This fixes current hardcoded list of 3 different pion particle types
in the initialization.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>