collisions: Fix checking of minimal time in collision_criteria_geometry()
It is needed to check both particles and not only one,
as otherwise we collide on the same time step twice with
the same particle.

While at it be more precise at removing the right id from the
list and add a bit of debug output.

Bump the version id for corrected physics.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>