ParticleType: isospin now more explicit property, charge added
ParticleType value 'isospin' is now handled identically with other properties
(meaning it must be explicitly set along with other particle properties).

The charge of the particle is also added as a ParticleType value and must be
set along with the other values.

In 'input-particles.cc' pdgcode is changed from float to int.

An example list of particles, 'particles.txt' is added to src folder.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>