print_measurements: Bugfix
As it is no longer certain that the first particle in the particlelist
remains active for the whole simulation, the time for output is calculated
from the last active particle. Since rhos are not yet active, the total
momentum of the system or the particle number do not stay constant.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>