Bug fixes and additional debug output
The initial position of resonance is now calculated properly in the
computational frame. Added energy and momentum conservation check after
the collision and some debug output on the new stuff.

Segfault issue still remains. Based on propagation data output
it seems that the new particle is not correctly assigned on particle list,
although all looks nice in collisions.cc.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>