box: Use iterators to go over the ParticleData vector array
This is necessary as elements might just disappear and hence
there is no longer a strict correspondence beween id of the particle
and the position in the particle vector.

Be careful to fix debug build too due to iterator change.

This is unrelated to a possible switch of set of the particle list.
Although it might make the switch useless.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>