OSCAR output function modified
write_oscar has been modified and overloaded. It now prints out
information for only one particle at a time and should be called separately
for each particle participating in the process. This is to generalize
the output for N->M processes, where N and M can have any values.

The function is overloaded to optionally take in two integers,
in addition to particle data and type. These integers
represent the number of particles in initial and final states of the
process (the above N and M). The first particle to be written in any process
should have these integers included in the function call. For the rest of the
particles in the same process, the simpler form of write_oscar should be used.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>