outputroutines: modify order and kick out total energy
The total energy doesn't change so no point in outputting it,
just output the small numerical difference.
This is anyway the check one wants to have a look at.

Have the timing output to be again the most right column,
to quickly check performance issues.

As added bonus the output now fits again in usual terminal width.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>