sphere: let the particles expand freely
Initializes a finite isotropic density profile that is going
to expand over time (For the moment beeing a step function).
This uses directly the "Box" the class too.

This is meant to stress test the virtual grid and in future
the use of stochastic verses geometrical collision criterion
in regions of different density.

The virtual grid is indeed different, that will have to move
to proper file both from box too.

This commit fixes refs #6.

TODO: fix cmake suckage, recompiling same objects for both targets.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>