bin: Add example loewe script
It is useful to know how mash can be run on loewe.
Also by git cloning it automaticaly ends up in the source.

Used for running 50 runs with:
for i in $(seq 50); do sbatch loewe_run.sh; done

Only the random number is set by cli all the other parameter
come from the params file. The strange naming of the output
directory is to find again the different variation of the
parameters. Also each output directory is unique.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>