particles.txt: Fermions added
Added proton, antiproton, neutron and delta resonance to particle list.
Since these are spin-1/2 particles, let's multiply all the particle spins
and isospins by 2 so we can still play with integers.

The change in the spin definition is applied also to resonance cross section
computation, along with the fact that there's now resonance with charge 2.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>