Particles: Add new overall class
It allows to more easily add and remove particles.
The type of a particle is now really easily accessed by id.
Use it all along as it simplifies all kind of function calls
and wired logic that was used for pdg lookups in resonances.
Hence reducing the code length there significantly.

It also gives easy access to the current number of particles
and the current computational time.

Change does_decay to check_decay and gob up logic of loop and
check, so that caller has less pedestrian work todo.
This is discussed with Jussi.
Change collision_time to const reference argument too.

Delete particles.h file, gob it up in Particles.h.
Cmake got confused by having both a lower and upercase include.
In the end some function might be better member function of that class.

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>