Particles: Improve class with specific methods
Provide direct accessor to ParticleData by specific iterators.
Disallow a copy of ParticleData as this is just big and
doesn't make sense.
Make sure to directly access the particle by it's id.

Make sure that ParticleType is passed down as a const reference,
so that it won't be changed.

Now at least mash runs fine for IC and segfaults only much later.

The crucial reference to find that we were infact accessing
copies of ParticleType and ParticleData is this post found by Jussi:
http://stackoverflow.com/questions/10472925/list-iterator-not-iterating-over-all-elements

Thanks: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>
Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>