Parameters: new class with key value pair for params file
This is general enough for now to be able to have different
parameter classes or subclasses.

Adapt paramreader to fill an vector of Parameters.
This is for the cleanup between box and sphere.

Note:
We can't pass const array as there ic no const vector iterator
before c11.

This commit refs #48

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>