param-reader: Add code to warn about unused params
While at it fix the code to really read in and assign all parameters.
The for loop was skipping entries, due to incrementing while removing
the entries.

Now we get again proper warnings if there is a wrong parameter.
Also this unentangles the param-reader from the box class only.

TODO: merge sphere + create sphere subclass

This commit refs #48

Signed-off-by: maximilian attems <attems@fias.uni-frankfurt.de>