Resonances: Lower mass limit revisited
There is now a more involved check for the minimum energy required
for resonance production. This is done in preparation for the situation
where resonances can actually form and decay to other resonances.

In addition, this change should prevent the occasional issues with
Delta resonances, where they could have too small mass for decays.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>