Decays: Bugfix and code beautification on decay_particles
The for-loop went through all decay modes even if already the first
one fulfilled the condition. As a result, particles with several
decay channels always used the last one in the list. This is fixed
by replacing the for-loop with a while-loop.

Also some more commentary has been added and some lines moved around
to make the code more readable.

Signed-off-by: Jussi Auvinen <auvinen@fias.uni-frankfurt.de>