"Test" for angular distributions
src/tests/angles_distribution.cc creates histograms in phi, costheta and
theta to check the distributions. It is not a test in the original
sense, because the Monte Carlo process can hardly be used for a fixed
comparison (as far as I am aware). But it can be used for manual
checking.