ctest: build tests with make all; deprecate check
The cmake style of tests is to enable building of tests via the
BUILD_TESTING option to cmake. Thus, if BUILD_TESTING is on, which it is
per default now, make all builds smash and all tests. This leads to more
thorough compile testing per default on development builds. If you want
short build cycles to work on one specific thing (e.g. because you're
doing test-driven development and only want to compile and run one test)
don't build the all target but a specific test target.

For releases of SMASH, the BUILD_TESTING flag should default to false.
Thus, dev builds default to testing and release builds default to
building only the application.

Reviewed-by: maximilian attems <attems@fias.uni-frankfurt.de>
Signed-off-by: Matthias Kretz <kretz@kde.org>