cmake: let make do the copy not cmake
If cmake does a file(COPY) then there are no dependency checks when make
is run. Therefore changes in the copied files are not carried to the
build dir. I changed it top instruct make to check for differences in
the files and copy from src to build dir if the file in src is newer.

Acked-by: maximilian attems <attems@fias.uni-frankfurt.de>
Signed-off-by: Matthias Kretz <kretz@kde.org>