testing: Import of Unit Test infrastructure from Vc
I looked through our options for a unit test infrastructure and nothing
is as simple to create tests as what I have in Vc. So copied my Vc unit
test header and stripped it down a bit.

The floating-point fuzzy compares calculates the ulp distance and
requires float compares. But with -Wfloat-equal that's noisy. So I
removed the compiler flag for the tests.

This refs #198

Signed-off-by: Matthias Kretz <kretz@kde.org>