tests: first unittest for Experiment
This test makes sure the factory method works for Box and Collider and
throws InvalidModusRequest for "Invalid".

To link the unit test against the relevant SMASH code I changed the
smash_asan target to build a static library first and then build
smash.cc against that lib. Then the lib is reused in the unittests.

Signed-off-by: Matthias Kretz <kretz@kde.org>