config: implemented taking key/values for bookkeeping
Now Configuration::take really removes the key/value pair. Next I
implemented a function that prints the remaining YAML. This should make
it easy to recognize unused configuration values.

Traversing through the YAML::Node hierarchy is very brittle: The copy
constructor of Node is *very* different to the copy assignment operator.
The former creates a handle to the node. The latter assigns the node on
the right hand side as value to the node on the left. The solution is to
use Node::reset instead of assignment.

Signed-off-by: Matthias Kretz <kretz@kde.org>