config: use the new Configuration class
main() creates the Configuration object and overwrites STEPS or MODUS.
Then the Experiment factory takes the object to determine the Modus. It
is passed on to the Experiment ctor which passes it further on to the
Modus ctor. Each then initialize their members from it. Finally main()
prints the remaining unused configuration.

Thus, now the initialization of the member variables happens from the
constructor directly. This implies we can now make some of these
variables "const".

I removed param-reader since it's now replaced by the YAML based
Configuration class.

This refs #76, #68, #187

Signed-off-by: Matthias Kretz <kretz@kde.org>