docs: unittest documentation
I added a little intro for using the unittest "framework" and documented
the available macros.

In addition doxygen now runs in the src directory - still outputting to
the build dir, though - in order to get the paths of the source files
relative to the src dir. This makes much more sense than showing the
path setup of my system in the public documentation.

Signed-off-by: Matthias Kretz <kretz@kde.org>