doc: improve layout and generate source browser
I changed the layout of the HTML output. I find this easier to work
with. Also I moved the detail description of classes and groups to the
top. These should be the most important parts of the documentation that
you'd want to see before anything else.

Also I disabled LaTeX output. For now there were no plans to use it.

Finally, I enabled source views throughout the documentation. This way
the doxygen output is very much geared towards code exploration.

Signed-off-by: Matthias Kretz <kretz@kde.org>