cmdline: parameters for particles/decaymodes overrides
Put particles and decaymodes contents into Configuration object. This
way it's very easy to override via different methods. (Sadly the
yaml-cpp output formatter creates a really ugly string in this case. I
hope this improves in a future version or we'll have to look for
alternatives.)

Particles is now a by-value member of Experiment and not recreated for
every event. This required a new Particles::reset() method which resets
id_max_ and clears the particle data map. But the particle types and
decay modes don't need to be recreated anymore.

This refs #258, #187, #68, #76

Signed-off-by: Matthias Kretz <kretz@kde.org>