Box: implement periodic boundary check as range algorithm
I generalized the BoxModus::boundary_condition function and made it
follow the standard STL pattern for a sequence modifying function.
Before, the function took a copy of a FourVector and returned another
new copy of a possibly modified FourVector. It additionally needed to
return a boolean flag and therefore had an output parameter.
Using the iterator algorithm pattern prevalent in the STL I rewrote it
to work on an arbitrary range and return the boolean with the return
value. This has the benefit of executing the necessary operations
directly on the FourVector without extra copies.

This refs #254

Signed-off-by: Matthias Kretz <kretz@kde.org>