VTK output: correct output path + generic interface
I created a new OutputInterface class that abstracts outputs into one
common interface. The first implementation is a port from the write_vtk
function. Since the output writer objects are now created in Experiment
it was easy to pass the right output path to the object. The object then
knows where to place the files.

This refs #226

Signed-off-by: Matthias Kretz <kretz@kde.org>