Tests: nucleus; angles
nucleus is tested for various properties. In TEST(center) and
TEST(center_hard_sphere), we test using the standard deviation of what
we know is the underlying distribution, failing when we are off more
than 3 sigma.
This refs #404
This refs #405

angles_zero tests if the x,y and z values of the angles add up to zero
in the long run. There is no actual test there, though; the values are
simply written in the end.
(This also refs #404)

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>