Nucleus: Bugfix in Woods-Saxon, test for WS.
distribution_nucleons() used the new rng::exponential(), but that
somehow broke the code. Now, we're back to -log(rng.uniform(0.0,1.0)),
which works.

There is a new TEST( woods_saxon ) now, which probes the distribution at
9 different points(*). To aid this, Nucleus has been extended to be able
to return the actual distribution from a function called woods_saxon()
which uses the current values of nuclear_radius() and diffusiveness_.

*: We check @f(x)/f(nuclear_radius())@, since for this we don't need to
know the normalization.