Woods-Saxon: New test to check/demonstrate centering
test woods-saxon demonstrates the difference between the direct
woods-saxon distribution and that when each nucleus is separately
centered. Thus, we can estimate the difference between a run with 1
testparticle and N testparticles (the direct distribution will for
N_runs -> infinity yield the same result independent of the number of
testparticles, while the centered one will differ).

SPOILER ALERT:

The centered distribution is a little more narrow than the direct
distribution.

This refs #404

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>