Random: Got rid of <double> specifier in canonical
The <double> template specifier in canonical is unnecessary now that the
declaration states double to be the default type.

Also, changed back the "auto" in the declaration of uniform_length in
boxmodus. This had been put in in the vain hope that the type of that
call might be fixed, but the static cast used now for the second
argument handles that.

This refs #69

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>