cmake: find and link against ROOT (optionally)
I rewrote FindROOT.cmake, making it much simpler and straightforward.
I adjusted the CMakeLists.txt to have a USE_ROOT option (defaults to
OFF). ROOT is not search if this is OFF, and REQUIRED if it's ON.

refs #443

Signed-off-by: Matthias Kretz <kretz@kde.org>

cmake: fix compilation with USE_ROOT=OFF

needed a corresponding #define and #ifdefs in experiment.cc
also needs to skip compilation of rootoutput.cc

refs #17

Signed-off-by: Matthias Kretz <kretz@kde.org>