tests: make configuration test independent of map order
the YAML classes internally use a map for storing the config tree. But
this implies that there is no guarantee on the order of config values.
This made the unit test very fragile. Now the test uses its own
config.yaml file and does not depend on a specific order anymore.

refs #366

Signed-off-by: Matthias Kretz <kretz@kde.org>