unittest: drop the need for std::function
Finally I realized how I can implement this without std::function.
Wasn't so hard in the end. :)

std::function was required because of the closure that contained the
try/catch block. Now this is done via a static member function which
knows what function to call through a template parameter. This
parameters type has a static member function test_function that contains
the test.

Why care? Because std::function makes it harder to spot the actual
problem when a test doesn't compile.

Signed-off-by: Matthias Kretz <kretz@kde.org>