PdgCode: Solve accessors with union
dump(), quarks() are now solved with a triple union:

the original bitfield is now in a struct digits_, which is in a union
with an integer dump_ and a struct chunks_.

chunks_ allows direct access to the quark digits and the excitation
digits (the 5th-7th digits from the end).

Please note that the order of the bitfield fields is
implementation-defined, thus a check for the implementation is added
with preprocessor macros.

Also, quarks() does not return 0xabc0 for quarks abc, but 0xabc
(trailing 0 is cut), and multiplett() returns a different bit mask now
than before. The documentation now says that multiplett() should only be
used for comparisons, not the value itself.

This refs #532

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>