Oscar: move code of initial/final state output
The non-member function is not required anymore as the only caller was
OscarOutput. I refactored the function such that the common part is in
write(Particles) and the specific line for initial/final is written with
a single fprintf statement.

At this point there may be an issue with writing to collision.dat,
because the non-member write_oscar function still opens the file with
every call, not using the already open file_ FilePtr.

Signed-off-by: Matthias Kretz <kretz@kde.org>