actions: vector<ActionPtr> as return value
Instead of using an in/out parameter for the std::vector<ActionPtr> we
now return the list of actions that find_possible_actions finds. In
order to append the list to the vector in experiment.cc we need an
append method for std::vector: We now have operator+= for
std::vector<ActionPtr> in action.h

Also I #if 0'ed the GridScatterFinder class. Still needs a comment that
this code is here for reference only.

Signed-off-by: Matthias Kretz <kretz@kde.org>