ParticleData: new ctor with PdgCode parameter
Allow ParticleData creation from a PdgCode. This is useful e.g. in
Action where the outgoing_particle list initially only contains the PDG
codes. Later the rest of the ParticleData is filled.

While at it I cleaned up the initialization of "improbable" values, so
that there's no need to duplicate them.

And a unit test for the new ctor.

Signed-off-by: Matthias Kretz <kretz@kde.org>