User Guide: Started documenting input options.
Doxygen can now create a rudimentary user guide. Use `make user` to
create.

Most of the content is proof-of-principle. To put something in the User
Guide, preceed the Doxygen block with "// !!USER:##SECTION##" (where
"##SECTION##" is replaced with the section this should be in; currently,
`input` and `Output` are recognized) and close that section with
"// !!/USER:##SECTION##".

If there is something that should go in the User Guide but not the
normal code documentation (`\page` markers and additional headers), put
it inside "\if user" ... "\endif" (these need to be inside the Doxygen
block).

Currently, Doxygen reads the sections from the src/include/*.h
files only and generates the user guide from all of these. The hierarchy
of the input file should be maintained by hand.

This refs #297

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>