Output: implement proper output calls from Box
Box keeps track of wrap-arounds via calls to
OutputInterface::write_interaction. These have the incoming and outgoing
particle as the particle before and after the coordinate transformation.

This change required the propagate method from Modus to take a list of
OutputInterface pointers as a third argument. To simplify use of the
type for this list of OutputInterface pointers I put it into the
forwarddeclarations header.

refs #564

Signed-off-by: Matthias Kretz <kretz@kde.org>