Action: new is_valid method / used in Experiment
up to now Action::perform would check that incoming_particles were still
valid. But since the creation of the ParticlesList would need the same
check I moved the code out into Action::is_valid. This is now called
from the loop in Experiment directly in order to skip the creation of
the ParticleList and the call to perform altogther. The perform method
itself retains the check via a call to is_valid, though, in order to be
on the safe side.

Signed-off-by: Matthias Kretz <kretz@kde.org>