Scatter/Resonance refactor
Removed lots more unnecessary map lookups. The ParticleData objects of
the outgoing particles are now set up in the outgoing_particles_ member
and only added to the global particle map at the end of the code. Also
ScatterAction::perform finally does the add to the Particles map, not
resonance_formation.

refs #444

Signed-off-by: Matthias Kretz <kretz@kde.org>