DecayAction: refactor calls to specific decays
The resonance_decay function really wasn't much more than a verbose
dispatch to one_to_two or one_to_three. And instead of a printed warning
I let the code throw an exception here now - because this is cleary an
error condition that is either fatal or must be handled higher up the
call stack.

Signed-off-by: Matthias Kretz <kretz@kde.org>