Clock: Rewinding of Clock not allowed
The timestep must never be negative, and neither can the time increments
in += operators: A std::range_error is thrown else. The time itself can
still be negative.

The test has been extended to check that, and also to check negative
times in multiple_is_in_next_tick.

This refs #401

Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>