includes: move all forward decls to single header
From the issue description:
The point here is that some forward decls are hard to get right and this
is boring and repetitive work. Especially if a declaration additionally
needs be hidden from doxygen... Including a single header for all
possible forward decls seems like a good idea to clean this up.

refs #236

Signed-off-by: Matthias Kretz <kretz@kde.org>