New DeformedNucleus class extends Nucleus
A deformednucleus class was added as a child of the nucleus class. The class
includes a deformed woods saxon sampling routine and a function to set specific
woods-saxon parameters for Uranium, Lead, Gold, and Copper (references included).
Outside the addition of the new class file, a "getter" method was added to the
Nucleus::diffusiveness private member for use in the child class, and the name
of Nucleus::distribution_nucleons was changed to Nucleus::distribute_nucleon.
This commit refs #324

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>