Clock: next_mutiple debugged.
This refs #697:

The old code checked the next multiple of the interval for each time
step; at very high time steps (counter_ = 6553630), this failed due to
floating point precision issues.

Now, we check what the next multiple is for the current_time and what it
is for next_time. If both are equal, then the next multiple is
considered not to be in the current time step. If they are not equal,
then the next multiple must be in the current time step.

Reported-by: Dmitry Oliinychenko <oliiny@fias.uni-frankfurt.de>
Signed-off-by: Bjørn Bäuchle <baeuchle@th.physik.uni-frankfurt.de>