Edit nucleusmodus for integrating deformations
Changes made to nucleus and nucleusmodus in order to
incorportate the deformednucleus into the workflow.
Includes new configuration tag for AUTOMATIC to automatically
set up the nucleus parameters.  Also adds setters and getters
for all nuclear variables (including saturation density, which
is now simply a normalization constant but is used in deformed
nucleus until a better sampling routing is picked).

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>