Initial boundaries for nuclei, shift update
Shift was made virtual to allow independence in
call at nucleusmodus.  The deformednucleus uses
a max radius to determine the boundaries of the shift
(i.e. the separation to generate on the z_axis). This
solution gives a boundary that is less clear cut as the
spherical nuclues, but my feeling is this is the most
time efficient way to do the shift.  Also, minor
aesthetics were altered.

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>