ROOT output: clean up of one-tree realization
Main problem of many trees in one root-file is that
with every new tree memory consumption grows. If number
of output moments is large enough it ends up in memory overflow.
With one tree memory consumption is limited and speed is
roughly the same. Viewing root-file, however, becomes less convenient.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>