Orientation added to shift
The deformed nucleus are each given boundaries according
to the max value of their most distant nucleon which is
used by the shift method to separate the two nuclei.
Also, the nucleus gets rotated in the shift method by
random angles.  Impact parameter is still in the x
direction, irrelevant of rotation.  As well, there is
a new method called manual_nucleus, which takes all the
parameters previously set in nucleus modus and sets them
inside the nucleus class (the workflow is: if automatic,
do determine_nucleus, else manual_nucleus).  There is
potential to add a way to manually decide about the
orientation angles, but this is not done yet.