BinaryOutput: pass FILE* instead of string option
The string option was a hack working around unique_ptrs not wanting to
be copied. In this case we can just pass FILE* to the write functions
instead. This is fine/safe because the caller to write still owns the
pointer and thus the raw pointer is a non-owning reference.

Signed-off-by: Matthias Kretz <kretz@kde.org>