Fixed bugs in frames calculations.
Made changes to  frames calculations and switched get_velocities return from vector to tuple. Still not completely functional, but the math for center of momentum and center of velocity have been checked now, and seem to work. The code does not like to run when using the fixed target calculation.

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>