Reference frame calculations work.
The reference frame calculations are working as desired, and the information has been put on the wiki.  The new changes are final math edits to get the equations correct, and a new feature which checks if velocities get too close to one and throws an exception if this is the case.

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>