userguide: improve structure and formatting
* I moved the documentation of configuration keys to the places in the
code where they are used. This should (hopefully) make it easier to keep
the documentation synchronized with the code.
* I fixed the hierarchy of \page s by using \subpage instead of \ref in
the right places.
* The order of parsing the sources by doxygen is important as it
determines how different \page sections that write to the same page are
concatenated. This needed adjusting in the developer docs where the .cc
files were read before the .h files.
* I added a \key alias for markup of config keys. I then made all the
config option pages consistent in their markup.
* I modified the integration of the config documentation into the
developer docs to remove most of the @\if user@ stuff.

refs #809

Signed-off-by: Matthias Kretz <kretz@kde.org>