userguide: simplified definition of userguide comments
I found a way to make the definition of user guide comments even
simpler: Now start a normal doxygen comment block with /** or /*!
(depending on whether you want to have auto-brief or not) and follow it
with \Userguide (there may be spaces in between, but no newlines). The
next */ closes the user guide comment section.

The "Userguide=" ALIAS is then used to make doxygen ignore the command
completely. Thus, it's only a marker for the "user guide extractor".

refs #809

Signed-off-by: Matthias Kretz <kretz@kde.org>