fix config.h usage and generation
Never generate files in the source directory! Generated files go to
CMAKE_(CURRENT_)BINARY_DIR. Subsequently do the includes correctly:
since config.h is not in the source tree there is no fixed relative
location to a source file in the source tree. Thus using #include with
"" is incorrect, it needs to use <>. The ${PROJECT_BINARY_DIR}/src
directory is in the include paths, so it's got to be <include/config.h>.
As this file was already getting generated, the configure_file call in
tests/CMakeLists.txt can go away.

refs #678, #819

Signed-off-by: Matthias Kretz <kretz@kde.org>