Moved rotate to 3vector, and documentation changes
Rotate is now a function of the threevector object. Also made compliance with doxygen comments (I think...).  Finally, there is now also a deformed nucleus test file, although it is rather useless at the moment do to me not understanding ulp. Last note is that the rotate method was not checked as extensively as before and also now includes the option for the third euler angle (called psi), which was not needed before. Although it wasn't checked as much, I think everything is in order.

Signed-off-by: Andy Goldschmidt <agoldschmidt11@gmail.com>