OSCAR: simplify implementation for format selection
It took a while until I realized that a function call can get rid of the
code duplication for the 1999/2013 format selection. The issue was that
for every possible configuration setup a different _type_ needs to be
deduced. You can't store types in variables... but you can pass them as
template parameters.
=> less code duplication & clearer logic for reading configuration
values

refs #837

Signed-off-by: Matthias Kretz <kretz@kde.org>