Configuration: new take function with fallback value
The new take overload requires a second argument which is returned from
the function if the requested key does not exist in the config file.
This makes it easier to implement config entries that are optional. At
the same time it makes the cast to the requested time more robust
because the function can infer the type from the type of the default
value.

Signed-off-by: Matthias Kretz <kretz@kde.org>