cmake: allow non-ASCII in particles.txt and decaymodes.txt
cmake's file(STRINGS ...) reads lines from a file - but only ASCII, it
simply discards everything else. That's bound to lead to surprises. So
instead the script now reads the complete file and converts it to a
cmake list by replacing \n with ; (to allow ; in lines those are
converted to \; first).

Signed-off-by: Matthias Kretz <kretz@kde.org>