EXPERIMENTAL: use UTF-8 greek symbols for particle names
The particle names are only used (and intended) for debug output. But
one important goal for debug output is making the output easy to parse
for humans, which can be vastly improved via aligning values in
different lines. To align values of different length, they have to be
padded with spaces. If the length of values is very different then this
can lead to huge horizontal gaps in the output. Therefore it is
desirable to have as short names as possible.

In addition to improved looks of the debug output, I find it nicer to
read anyway.

This commit is marked as experimental, because it is not clear how well
the utf-8 support works in all the terminals that we want to support.

refs #853

Signed-off-by: Matthias Kretz <kretz@kde.org>